STNDP-134 Add invite collaborators setting
